backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
pydantic
openai
python-dotenv
